content/media/video/test/test_seek1.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: seek test 1</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<video id='v'
       src='seek.ogg'
       onended='return playbackEnded();'
       onloadedmetadata='return startTest();'
       onseeking='return seekStarted();'
       onseeked='return seekEnded();'></video>
<pre id="test">
<script class="testbody" type="text/javascript">
var startPassed = false;
var endPassed = false;
var seekFlagStart = false;
var seekFlagEnd = false;
var timeStart;
var readonly = true;
ok(!document.getElementById('v').seeking, "seeking should default to false");
try {
  v1.seeking = 1;
  readonly = false;
}
catch(e) {
  readonly = true;
}
ok(readonly, "seeking should be readonly");

function startTest() {
  var v = document.getElementById('v');
  v.play();
  v.currentTime=2;
}

function seekStarted() {
  var v = document.getElementById('v');
  startPassed = true;
  seekFlagStart = v.seeking;
  return false;
}

function seekEnded() {
  var v = document.getElementById('v');
  var t = v.currentTime;
  ok(t >= 2 && t <= 3, "Video currentTime is incorrect: " + t);
  endPassed = true;
  seekFlagEnd = v.seeking;
  timeStart = new Date();
  return false;
}

function playbackEnded() {
  var t = (new Date()) - timeStart;
  ok(startPassed, "Video did not send seeking event");
  ok(endPassed, "Video did not send seeked event");
  ok(t > 0 && t <= 2500, "Video did not seek to correct time: " + t);
  ok(seekFlagStart, "Seek flag not set on seek start");
  ok(!seekFlagEnd, "Seek flag set on seek end");
  SimpleTest.finish();
  return false;
}

setTimeout(function () {
             ok(false, "Test timed out");
             SimpleTest.finish();
           }, 30000); 

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
